Simplify Editor Modal
